requirements.txt
python-telegram-bot
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
